1) сконверти в CSV пока без даты/времени. Посмотрел в Excel вроде норм. 2) Есть проблема в файлах которые записывает DExS от 30% до 50% файлов имеют смещение на 2 байта почему - пока не понятно, дата в этих файлах будет некорректна

+++ src/index.ts
@@ -3,7 +3,7 @@ import { Settings } from "./utils/config";
 
 console.log('Start DUMB2CSV');
 
-const src: Uint8Array = new Uint8Array(fs.readFileSync('./data/22162724.DMP'));
+const src: Uint16Array = new Uint16Array(fs.readFileSync('./data/22162724.DMP'));
 
 class TDump{
   T1000Hz: number;//u16
